+++ CRM/Mautic/Upgrader.php
@@ -95,7 +95,7 @@ class CRM_Mautic_Upgrader extends CRM_Extension_Upgrader_Base {
 
 
   /**
-   * Insert Civirules trigger for WebHook.
+   * Insert Civirules trigger for Webhook.
    *
    * @return TRUE on success
    * @throws Exception
@@ -143,7 +143,7 @@ class CRM_Mautic_Upgrader extends CRM_Extension_Upgrader_Base {
     $this->executeCustomDataFileByAbsPath($file);
     return TRUE;
   }
-  
+
   public function upgrade_4207() {
     $this->ctx->log->info('Action to add contact to Event/Segment.');
     $this->enableCiviRules();
@@ -155,6 +155,6 @@ class CRM_Mautic_Upgrader extends CRM_Extension_Upgrader_Base {
     $this->enableCiviRules();
     return TRUE;
   }
-  
+
 
 }